Classify observations

Classify observations according to the maximum a posterior probabilites.

classify(x, theta)

Arguments

x
Either a matrix of A) observations where rows corresponds to obsercations and columns to dimensions or B) class probabilities where rows correspond to obsevations and columns to components.

theta
A list of parameters for the full model as described in rtheta. If theta is supplied, x are assumed to be observations (A). If theta is missing, x are assumed to be probabilites (B).

Value

A integer vector of class numbers with length equal to the number of rows in x.

See also

Examples

# Classify using probabilites (usually returned from get.prob)
probs <- matrix(runif(75), 25, 3)
classify(probs)

#>  [1] 3 1 3 3 1 3 3 3 1 1 1 3 1 2 2 3 1 3 3 2 2 1 2 3 3

# Classify using a matrix of observations and theta
theta <- rtheta(d = 4, m = 3)
u <- SimulateGMCMData(n = 20, theta = theta)$u
classify(x = u, theta = theta)

#>  [1] 3 2 2 3 2 2 2 1 2 3 3 2 2 3 2 3 3 1 2 2