Convert between parameter formats

These functions converts/coerces the parameters between the general Gaussian mixture (copula) model and the special GMCM. Most functions of the GMCM packages use the theta format described in rtheta.

full2meta(theta)

meta2full(par, d)

Arguments

theta
A list of parameters for the full model. Formatted as described in rtheta.

par
A vector of length 4 where par[1] is the probability of coming from the first component, par[2] is the mean value, par[3] is the standard deviation, and par[4] is the correlation of the reproducible component.

d
An integer giving the dimension of the mixture distribution.

Value

full2meta returns a numeric vector of length 4 formatted as par.

meta2full returns a formatted 'theta' list of parameters as described by rtheta.

Details

If a theta is supplied which is not on the form of Li et. al. (2011) the output is coerced by simply picking the first element of the second component mean vector as mean, the square roof of the first diagonal entry of the second component covariance matrix as standard deviation, and first off-diagonal entry as correlation (properly scaled).

References

Li, Q., Brown, J. B. J. B., Huang, H., & Bickel, P. J. (2011). Measuring reproducibility of high-throughput experiments. The Annals of Applied Statistics, 5(3), 1752-1779. doi:10.1214/11-AOAS466

Tewari, A., Giering, M., & Raghunathan, A. (2011). Parametric Characterization of Multimodal Distributions with Non-gaussian Modes. IEEE 11th International Conference on Data Mining Workshops, 2011, 286-292. doi:10.1109/ICDMW.2011.135

See also

Examples

theta <- GMCM:::rtheta(m = 2, d = 2)
print(par <- full2meta(theta))

#>        pie1          mu       sigma         rho 
#>  0.09074447 10.27417079  2.40060596 -0.00726274 

print(theta.special.case <- meta2full(par, d = 2))

#> theta object with d = 2 dimensions and m = 2 components:
#> 
#> $pie
#>       pie1       pie2 
#> 0.09074447 0.90925553 
#> 
#> $mu
#> $mu$comp1
#> [1] 0 0
#> 
#> $mu$comp2
#>       mu       mu 
#> 10.27417 10.27417 
#> 
#> 
#> $sigma
#> $sigma$comp1
#>      [,1] [,2]
#> [1,]    1    0
#> [2,]    0    1
#> 
#> $sigma$comp2
#>             [,1]        [,2]
#> [1,]  5.76290897 -0.04185451
#> [2,] -0.04185451  5.76290897
#> 
#>