Logit and inverse logit transforms

The logit transformation (i.e. the log of the odds) and its inverse (also called expit).

inv.logit(a)

logit(p)

Arguments

a
A vector of real values.

p
A vector of probabilities.

Value

inv.logit returns a vector of the same length as a of the inverse logit transformed values. This function is also known as the expit-function.

logit returns a vector of the same length as p with the log odds of p.

See also

Examples

p <- runif(100)
print(a <- GMCM:::logit(p))

#>   [1]  1.17124004 -2.73307596  0.13589128  0.22656430  1.73226516 -0.52895520
#>   [7] -3.39776953 -0.92333308  0.53148654  0.31963554  1.70903332  2.89205423
#>  [13]  0.89173935  3.56262043  2.51504408  1.28588764 -0.11846614  2.70628791
#>  [19] -0.18067360  1.84397139  1.41233384 -1.13523080 -0.29661387  2.38224288
#>  [25]  2.48892173 -4.71837983  3.96134363 -0.36109652 -0.91885810  0.23026889
#>  [31] -1.82905135  0.39944136  2.27459479 -0.36245971  1.17291355 -0.84549689
#>  [37]  1.17520691  0.20845762 -1.60525954  0.42354994  0.26596496  0.55775771
#>  [43]  1.77845340  0.48369242 -0.58797794  1.98873009  0.54505126 -1.85354961
#>  [49] -4.56834790  0.05953090  2.01778537 -0.72098594 -2.41264624  3.71719947
#>  [55] -2.81819807  1.05452860  2.23904432 -0.88386826  1.20483362 -0.47596303
#>  [61] -0.92488928 -1.27201961 -0.67622820 -1.64434775  1.48185158  2.99505022
#>  [67]  1.83451335  1.81598767 -0.08303569 -1.76620162 -1.09114118 -0.68235429
#>  [73] -0.64222593 -1.18857857  0.69832310 -0.74342810 -0.29970802 -0.92944502
#>  [79] -1.65686514  2.38471191 -1.52888197 -0.40650082 -0.03604368  1.65464709
#>  [85] -0.01713842 -0.59370940 -0.80735636 -0.81797730 -0.29034178  0.92755510
#>  [91]  0.30295922  2.46049503  0.81189883  0.26063944  3.00783397 -2.26731363
#>  [97]  5.21053374  1.02812470  0.84357037  2.07038493

p - GMCM:::inv.logit(a)

#>   [1]  0.000000e+00 -6.938894e-18  0.000000e+00  0.000000e+00  0.000000e+00
#>   [6]  0.000000e+00  6.938894e-18  0.000000e+00  0.000000e+00  0.000000e+00
#>  [11]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [16]  0.000000e+00  0.000000e+00  0.000000e+00 -5.551115e-17  0.000000e+00
#>  [21]  0.000000e+00  0.000000e+00 -5.551115e-17  0.000000e+00  0.000000e+00
#>  [26] -1.734723e-18  0.000000e+00  5.551115e-17  0.000000e+00  0.000000e+00
#>  [31] -2.775558e-17  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [36]  0.000000e+00  1.110223e-16  1.110223e-16  0.000000e+00  0.000000e+00
#>  [41]  0.000000e+00 -1.110223e-16  0.000000e+00 -1.110223e-16 -5.551115e-17
#>  [46]  0.000000e+00  0.000000e+00  0.000000e+00 -1.734723e-18  0.000000e+00
#>  [51]  1.110223e-16  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [56]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [61]  0.000000e+00  0.000000e+00  0.000000e+00  2.775558e-17  0.000000e+00
#>  [66]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [71]  0.000000e+00  0.000000e+00  5.551115e-17 -2.775558e-17  0.000000e+00
#>  [76]  0.000000e+00  0.000000e+00  0.000000e+00  2.775558e-17  0.000000e+00
#>  [81]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  5.551115e-17
#>  [86] -5.551115e-17  0.000000e+00  0.000000e+00  5.551115e-17  0.000000e+00
#>  [91]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [96]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00 -1.110223e-16