Transformation of the correlation to real line and its inverse

A transformation of the correlation coefficient into the real line and the corresponding inverse. The transform is a translation and scaling of \(\rho\) from the interval \((-1/(d-1), 1)\) to \((0, 1)\) followed by a logit transformation to the whole real line.

inv.rho.transform(a, d)

rho.transform(rho, d)

Arguments

a
A real number.

d
The dimension of the space.

rho
A correlation coefficient between -1/(d-1) and 1.

Value

inv.rho.transform returns a vector of the inversely transformed values with the same length as a.

rho.transform returns a vector of the transformed values with the same length as rho or a.

See also

Examples

d <- 4
rho <- runif(100, -1/(d-1), 1)
a <- GMCM:::rho.transform(rho, d)
rho - GMCM:::inv.rho.transform(a, d)

#>   [1] -2.081668e-17  0.000000e+00 -5.551115e-17 -1.110223e-16  0.000000e+00
#>   [6]  0.000000e+00  1.387779e-17  2.775558e-17  1.110223e-16  1.110223e-16
#>  [11]  0.000000e+00  2.081668e-17  0.000000e+00  0.000000e+00 -1.110223e-16
#>  [16] -2.081668e-17 -6.938894e-17  8.326673e-17 -2.775558e-17  0.000000e+00
#>  [21] -1.110223e-16  5.551115e-17  0.000000e+00  0.000000e+00  2.775558e-17
#>  [26]  2.775558e-17  1.110223e-16  2.775558e-17  5.551115e-17  2.775558e-17
#>  [31]  0.000000e+00 -1.110223e-16  0.000000e+00 -1.110223e-16  0.000000e+00
#>  [36]  1.110223e-16 -1.110223e-16 -5.551115e-17  0.000000e+00 -1.110223e-16
#>  [41] -2.775558e-17  0.000000e+00  1.110223e-16 -5.551115e-17 -1.387779e-17
#>  [46]  0.000000e+00 -1.110223e-16  0.000000e+00  0.000000e+00  0.000000e+00
#>  [51]  0.000000e+00  0.000000e+00 -1.110223e-16  0.000000e+00  0.000000e+00
#>  [56]  1.110223e-16  0.000000e+00 -5.551115e-17  0.000000e+00 -1.110223e-16
#>  [61]  0.000000e+00 -1.387779e-17  0.000000e+00 -1.110223e-16  1.110223e-16
#>  [66]  0.000000e+00  0.000000e+00  0.000000e+00  1.110223e-16 -2.775558e-17
#>  [71]  0.000000e+00 -1.110223e-16  0.000000e+00 -1.110223e-16  2.081668e-17
#>  [76]  0.000000e+00  5.551115e-17 -1.110223e-16  2.220446e-16  0.000000e+00
#>  [81]  1.110223e-16 -2.081668e-17  0.000000e+00  8.326673e-17  0.000000e+00
#>  [86] -5.551115e-17 -1.110223e-16  0.000000e+00  2.775558e-17  1.110223e-16
#>  [91]  0.000000e+00 -1.110223e-16  0.000000e+00 -1.665335e-16  1.387779e-17
#>  [96]  0.000000e+00 -9.367507e-17  1.110223e-16 -1.110223e-16  2.775558e-17