Reparametrization of GMCM parameters

These functions map the four GMCM parameters in the model of Li et. al. (2011) and Tewari et. al. (2011) onto the real line and back. The mixture proportion is logit transformed. The mean and standard deviation are log transformed. The correlation is translated and scaled to the interval (0,1) and logit transformed by rho.transform.

inv.tt(par, d, positive.rho)

tt(tpar, d, positive.rho)

Arguments

par
A vector of length 4 where par[1] is the mixture proportion, tpar[2] the mean, tpar[3] the standard deviation, and tpar[4] the correlation.

d
The dimension of the space.

positive.rho
is logical. If TRUE, the correlation is transformed by a simple logit transformation. If FALSE the rho.transform is used.

tpar
A vector of length 4 of the transformed parameter values where tpar[1] corresponds to the mixture proportion, tpar[2] the mean, tpar[3] the standard deviation, and tpar[4] the correlation.

Value

inv.tt returns tpar as described above.

A numeric vector of the transformed or inversely transformed values of length 4.

tt returns par as described above.

Details

The functions are used only in the wrapper to optim when the GMCM log-likelihood is optimized.

par[1] should be between 0 and 1. par[2] and par[3] should be non-negative. If positive.rho is FALSE, par[4] should be between \(-1/(d-1)\) and 1. Otherwise, positive.rho should be between 0 and 1.

References

Li, Q., Brown, J. B. J. B., Huang, H., & Bickel, P. J. (2011). Measuring reproducibility of high-throughput experiments. The Annals of Applied Statistics, 5(3), 1752-1779. doi:10.1214/11-AOAS466

Tewari, A., Giering, M. J., & Raghunathan, A. (2011). Parametric Characterization of Multimodal Distributions with Non-gaussian Modes. 2011 IEEE 11th International Conference on Data Mining Workshops, 286-292. doi:10.1109/ICDMW.2011.135

Examples

par <- c(pie1 = 0.3, mu = 2, sigma = 0.5, rho = 0.8)
tpar <- GMCM:::inv.tt(par, d = 3, positive.rho = FALSE)
GMCM:::tt(tpar, d = 3, positive.rho = FALSE)

#> [1] 0.3 2.0 0.5 0.8