Goodness of fit for the general GMCM

Compute goodness of fit as described in AIC. The number of parameters used correspond to the number of variables free to vary in the general model.

goodness.of.fit(theta, u, method = c("AIC", "BIC"), k = 2)

Arguments

theta
A list of parameters as defined in rtheta. For t this function, it will usually be the output of fit.full.GMCM.

u
An n by d matrix of marginally uniform observations. Rows corresponds to observations and columns to the dimensions of the variables. I.e. these are often ranked and scaled test statistics or other observations.

method
A character of length 1 which specifies the goodness of fit to compute. Default is "AIC". "BIC" is also a option.

k
A integer specifying the default used constant "k" in AIC. See AIC.

Value

A single number giving the goodness of fit as requested.

Examples

set.seed(1)
data(u133VsExon)
u <- Uhat(u133VsExon[sample(19577, 2000), ])  # Subset for faster fitting
theta1 <- fit.full.GMCM(u, m = 2, method = "L-BFGS")

#> iter   10 value -516.933962
#> iter   20 value -517.631433
#> iter   30 value -517.648687
#> final  value -517.648688 
#> converged

theta2 <- fit.full.GMCM(u, m = 3, method = "L-BFGS")

#> iter   10 value -517.850361
#> iter   20 value -519.842937
#> iter   30 value -521.666691
#> iter   40 value -524.466780
#> iter   50 value -524.973285
#> iter   60 value -525.118018
#> iter   70 value -525.150249
#> iter   80 value -525.187378
#> iter   90 value -525.247141
#> iter  100 value -525.330933
#> iter  110 value -525.368686
#> iter  120 value -525.371502
#> final  value -525.372367 
#> converged

# AIC
goodness.of.fit(theta1, u)

#> [1] -1019.297

goodness.of.fit(theta2, u)

#> [1] -1022.745

# BIC
goodness.of.fit(theta1, u, method = "BIC")

#> [1] -974.4902

goodness.of.fit(theta2, u, method = "BIC")

#> [1] -944.3321