Logit and inverse logit transforms

The logit transformation (i.e. the log of the odds) and its inverse (also called expit).

inv.logit(a)

logit(p)

Arguments

a
A vector of real values.

p
A vector of probabilities.

Value

inv.logit returns a vector of the same length as a of the inverse logit transformed values. This function is also known as the expit-function.

logit returns a vector of the same length as p with the log odds of p.

See also

Examples

p <- runif(100)
print(a <- GMCM:::logit(p))

#>   [1]  1.13786696  0.45159318  1.34843411  3.05830313 -2.27704891  1.12454434
#>   [7]  1.67604059  2.43090981  0.71019270  1.78181084  1.42359530 -0.99673087
#>  [13]  0.25770617 -3.50651795  0.26247473  4.58615097  3.48919447  1.52405316
#>  [19]  1.55402498 -1.97885280  3.29735477 -0.08866464 -0.15000137  3.25997158
#>  [25]  2.62817155 -2.41954846  0.98772099  0.27363579 -0.72256612 -2.99459435
#>  [31] -0.49313496 -0.88889016  1.37742294 -2.87002023  1.84186810  0.37244837
#>  [37]  2.20935721 -2.12861571 -2.66193185 -2.05291494 -0.23698314 -0.39680734
#>  [43]  0.61808623  4.21733494 -2.88267142  0.55187074  0.72412460  0.86341640
#>  [49] -1.78551632  0.34949230  0.88862088 -1.25924154 -0.32865296 -1.57573205
#>  [55]  2.28425900  5.12254747  0.30020845  0.42253173 -1.09777010  2.83059633
#>  [61]  2.28690912 -2.04441568 -0.36658189 -2.18205386  0.30549082  1.27715140
#>  [67] -1.39214987 -1.56250387 -3.24893650 -0.46534148 -2.45081525 -0.70333288
#>  [73]  1.65200080  0.01450171 -0.89923114  1.88484103 -2.85109199  0.13986157
#>  [79]  0.78071002 -0.56137145  0.08456287  1.95339384  2.75143398 -1.21347306
#>  [85] -0.31278124  0.92277872 -1.23848881 -1.07122907 -1.55610267  0.25577017
#>  [91] -0.63601735  0.85440502 -1.22736522  0.55155255 -0.56362849 -2.44088447
#>  [97]  1.71871785  0.30708104  0.80944801  0.95471312

p - GMCM:::inv.logit(a)

#>   [1]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00 -1.387779e-17
#>   [6]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [11]  0.000000e+00  0.000000e+00  0.000000e+00 -3.469447e-18  0.000000e+00
#>  [16]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [21]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [26]  1.387779e-17  0.000000e+00  0.000000e+00  0.000000e+00  6.938894e-18
#>  [31]  0.000000e+00  0.000000e+00  0.000000e+00  6.938894e-18  0.000000e+00
#>  [36]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [41]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  6.938894e-18
#>  [46]  1.110223e-16  0.000000e+00  0.000000e+00  0.000000e+00  1.110223e-16
#>  [51]  0.000000e+00  2.775558e-17  0.000000e+00  2.775558e-17  0.000000e+00
#>  [56]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [61]  0.000000e+00 -1.387779e-17  0.000000e+00  0.000000e+00  0.000000e+00
#>  [66]  1.110223e-16  0.000000e+00  2.775558e-17 -6.938894e-18  0.000000e+00
#>  [71] -1.387779e-17  0.000000e+00  0.000000e+00  0.000000e+00  5.551115e-17
#>  [76]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  5.551115e-17
#>  [81]  0.000000e+00 -1.110223e-16  1.110223e-16  0.000000e+00  5.551115e-17
#>  [86]  0.000000e+00 -2.775558e-17 -5.551115e-17  0.000000e+00  0.000000e+00
#>  [91]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [96]  1.387779e-17  0.000000e+00  1.110223e-16  0.000000e+00  0.000000e+00