Logit and inverse logit transforms

The logit transformation (i.e. the log of the odds) and its inverse (also called expit).

inv.logit(a)

logit(p)

Arguments

a
A vector of real values.

p
A vector of probabilities.

Value

inv.logit returns a vector of the same length as a of the inverse logit transformed values. This function is also known as the expit-function.

logit returns a vector of the same length as p with the log odds of p.

See also

Author

Anders Ellern Bilgrau <anders.ellern.bilgrau@gmail.com>

Examples

p <- runif(100)
print(a <- GMCM:::logit(p))

#>   [1] -1.37758753 -0.54518827 -2.59754610  0.50579740 -0.62729149  8.51621969
#>   [7]  1.42021037  0.59429078  1.73875409  2.99463153  3.04716722 -2.34245547
#>  [13]  2.80311255  2.30030444 -2.17421443 -0.45961335  0.80394612 -1.01003273
#>  [19] -1.43410176 -0.42265454 -0.30217263 -0.56257868 -1.77162457  0.41140760
#>  [25] -0.59403002  0.96752286  0.64589201 -0.07319073  0.14203067  0.34615293
#>  [31]  0.98419031  0.76305483  0.76616937  0.78278073  0.06353915  1.48668599
#>  [37] -1.54364034  0.11564227  1.55797768  0.07960936 -0.83572676  0.33589983
#>  [43] -0.60612849  0.40336187  2.18431340  0.18774704  1.30112340 -0.81268369
#>  [49] -0.67164004  0.94428307 -3.40413800 -1.31505630  0.58027572 -2.78957544
#>  [55] -1.28732370 -0.98179283  2.90686380 -3.48184590 -1.56722935  0.45316224
#>  [61] -0.56861741  2.62911450 -0.88801009 -0.85543981 -0.39507228  1.17605460
#>  [67]  0.63552215 -1.82276398  2.78297674  0.12381124 -0.68019807  1.32930349
#>  [73] -0.97156854  0.91149072  1.21454506  1.84280422  0.87224956  1.67561060
#>  [79] -0.93506324  0.54284806  1.14930506 -1.42262387  0.93620897  0.55277950
#>  [85]  0.66476490 -0.68506644  1.73949350  2.41762371  0.56623559  1.60362278
#>  [91]  2.04619891  0.92931918 -2.12872047  1.75265709 -0.93709934  2.36454748
#>  [97] -0.04229209 -0.54643567 -1.31210549  4.52939595

p - GMCM:::inv.logit(a)

#>   [1] -2.775558e-17  0.000000e+00  1.387779e-17  0.000000e+00 -5.551115e-17
#>   [6]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [11]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  1.387779e-17
#>  [16]  0.000000e+00  0.000000e+00  0.000000e+00  2.775558e-17  0.000000e+00
#>  [21]  0.000000e+00  5.551115e-17  2.775558e-17 -1.110223e-16  0.000000e+00
#>  [26]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00 -1.110223e-16
#>  [31]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [36]  0.000000e+00  2.775558e-17  0.000000e+00  0.000000e+00  0.000000e+00
#>  [41]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [46]  0.000000e+00 -1.110223e-16  0.000000e+00  0.000000e+00  0.000000e+00
#>  [51] -6.938894e-18  0.000000e+00  0.000000e+00 -6.938894e-18  0.000000e+00
#>  [56]  0.000000e+00  0.000000e+00  1.040834e-17  2.775558e-17  0.000000e+00
#>  [61]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [66] -1.110223e-16  0.000000e+00  2.775558e-17  0.000000e+00  0.000000e+00
#>  [71]  0.000000e+00  1.110223e-16  0.000000e+00  0.000000e+00  1.110223e-16
#>  [76]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [81]  1.110223e-16  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [86]  5.551115e-17  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [91]  0.000000e+00  0.000000e+00  1.387779e-17  0.000000e+00  0.000000e+00
#>  [96]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00